Improve micro typography
- import package 'microtype'
- introduce macro for "Netz39"
- replace all occurences of "Netz39 e.V." with new macro
- adapt horizontal whitespace and kerning for new macro \netzEV

The German recommendations for abbreviations recommend setting a thin
space in between parts of an abbreviation. See for example:

- http://faql.de/typographie.html#abk
- https://de.wikipedia.org/wiki/Schmales_Leerzeichen

With the font used (lmodern) the '\,' or 1/6 em wide whitespace is a good
start, but visually still a little wide, so a space slightly less than
an "Achtelgeviert" is used.

Additionally the kerning between the 'V' and the following dot was
improved. That dot stood out far too much to the right. Current state is
result of comparision with typographically well looking examples
I got from a professional typo expert on another channel.